--- a/internal/conv/string_test.go
+++ b/internal/conv/string_test.go
@@ -52,7 +52,7 @@ func (s *StringSuite) TestUnsafeBytesToStr() {
 }
 
 func BenchmarkUnsafeStrToBytes(b *testing.B) {
-	for i := 0; i < b.N; i++ {
+	for i := 0; b.Loop(); i++ {
 		UnsafeStrToBytes(strconv.Itoa(i))
 	}
 }